#8 - Separated the platform code into partial classes; Added no-op for FrameworkInfo on netstandard targets.
